refactor: 로또 클래스 위치 이동
- 입력 스토어 인스턴스 생성 방식 통일
