+++ src/models/index.js
@@ -1,4 +1,4 @@
-import Lotto from '../Lotto.js';
+import Lotto from './Lotto.js';
 import LottoChecker from './LottoChecker.js';
 import LottoResult from './LottoResult.js';
